Refactor code to put some duplicate code in helper
